Add files via upload
Crafting Sign-Up Flow
A sign-up flow that met the needs of the users. It is a simple, easy to understand and visually appealing design. The sign up process can be a boring process for the users. Therefore, it is important to keep it as simple and short as possible, also adding an option of sign up through Google and social media can make the process quick - avoiding drop offs.