Add files via upload
My next project was to design a restaurant's digital menu. It was an exciting challenge to blend the art of visual design with the practicality of presenting a diverse menu in a user-friendly manner. Even though now everyone is familiar with the digital world, still some don't prefer to use digital menu's.